added options for tr, fixed ticket release controller post

+++ pkg/database/factory/ticket_release_factory.go
@@ -4,7 +4,7 @@ import (
 	"github.com/DowLucas/gin-ticket-release/pkg/models"
 )
 
-func NewTicketRelease(eventID int, open, close uint, hasAllocatedTickets bool, ticketReleaseMethodDetailID uint) *models.TicketRelease {
+func NewTicketRelease(eventID int, open, close int64, hasAllocatedTickets bool, ticketReleaseMethodDetailID uint) *models.TicketRelease {
 	return &models.TicketRelease{
 		EventID:                     eventID,
 		Open:                        open,
