+++ pkg/database/factory/ticket_release_method_detail_factory.go
@@ -4,7 +4,7 @@ import (
 	"github.com/DowLucas/gin-ticket-release/pkg/models"
 )
 
-func NewTicketReleaseMethodDetail(maxTicketsPerUser uint, notificationMethod, cancellationPolicy string, openWindowDuration uint, ticketReleaseMethodID uint) *models.TicketReleaseMethodDetail {
+func NewTicketReleaseMethodDetail(maxTicketsPerUser uint, notificationMethod, cancellationPolicy string, openWindowDuration int64, ticketReleaseMethodID uint) *models.TicketReleaseMethodDetail {
 	return &models.TicketReleaseMethodDetail{
 		MaxTicketsPerUser:     maxTicketsPerUser,
 		NotificationMethod:    notificationMethod,
